Updating

+++ psi4/src/psi4/cctransort/memcheck.cc
@@ -3,35 +3,38 @@
  *
  * Psi4: an open-source quantum chemistry software package
  *
- * Copyright (c) 2007-2017 The Psi4 Developers.
+ * Copyright (c) 2007-2018 The Psi4 Developers.
  *
  * The copyrights for code used from other parties are included in
  * the corresponding files.
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
+ * This file is part of Psi4.
  *
- * This program is distributed in the hope that it will be useful,
+ * Psi4 is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * Psi4 is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with Psi4; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  *
  * @END LICENSE
  */
 
 #include "psi4/libdpd/dpd.h"
+#include "psi4/libpsi4util/PsiOutStream.h"
+
 
 namespace psi { namespace cctransort {
 
 void memcheck(int reference)
 {
-  unsigned long int irrep_size, size;
+  size_t irrep_size, size;
   dpdbuf4 Z;
 
   outfile->Printf( "\n");
@@ -40,7 +43,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 5, 5, 5, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <ab|cd> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -52,7 +55,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 10, 5, 10, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <ia|bc> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -64,7 +67,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 0, 5, 0, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of tijab amplitudes:  %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -79,7 +82,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 5, 5, 5, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <ab|cd> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -91,7 +94,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 10, 5, 10, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <ia|bc> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -103,7 +106,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 0, 5, 0, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of tIjAb amplitudes:  %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -117,7 +120,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 7, 7, 7, 7, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <AB|CD> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -128,7 +131,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 17, 17, 17, 17, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <ab|cd> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -139,7 +142,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 28, 28, 28, 28, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <Ab|Cd> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -151,7 +154,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 20, 5, 20, 5, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <IA|BC> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -163,7 +166,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 30, 15, 30, 15, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <ia|bc> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -175,7 +178,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 24, 28, 24, 28, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <Ia|Bc> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -187,7 +190,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 27, 29, 27, 29, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of <iA|bC> integrals: %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
@@ -199,7 +202,7 @@ void memcheck(int reference)
     global_dpd_->buf4_init(&Z, 99, 0, 22, 28, 22, 28, 0, "Just a template");
     size = 0;
     for(int h=0; h < Z.params->nirreps; h++) {
-      irrep_size = (unsigned long int) Z.params->rowtot[h] * Z.params->coltot[h];
+      irrep_size = (size_t) Z.params->rowtot[h] * Z.params->coltot[h];
       size += irrep_size;
       outfile->Printf( "\tSize of irrep %1lu of tIjAb amplitudes:  %10.3lf (MW) / %10.3lf (MB)\n",
               h, irrep_size/1e6, (irrep_size/1e6)*sizeof(double));
