+++ psi4/src/psi4/detci/s1v.cc
@@ -3,23 +3,24 @@
  *
  * Psi4: an open-source quantum chemistry software package
  *
- * Copyright (c) 2007-2017 The Psi4 Developers.
+ * Copyright (c) 2007-2018 The Psi4 Developers.
  *
  * The copyrights for code used from other parties are included in
  * the corresponding files.
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
+ * This file is part of Psi4.
  *
- * This program is distributed in the hope that it will be useful,
+ * Psi4 is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * Psi4 is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with Psi4; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  *
  * @END LICENSE
@@ -88,9 +89,9 @@ void s1_block_vfci(struct stringwr **alplist, struct stringwr **betlist,
                    int nlists, int nas, int nbs, int Ib_list, int Jb_list,
                    int Jb_list_nbs) {
   struct stringwr *Ib, *Kb;
-  unsigned int Ia_idx, Ib_idx, Kb_idx, Jb_idx;
-  unsigned int Ibcnt, Kbcnt, Kb_list, Ib_ex, Kb_ex;
-  unsigned int *Ibridx, *Kbridx;
+  size_t Ia_idx, Ib_idx, Kb_idx, Jb_idx;
+  size_t Ibcnt, Kbcnt, Kb_list, Ib_ex, Kb_ex;
+  size_t *Ibridx, *Kbridx;
   int *Ibij, *Kbij;
   signed char *Ibsgn, *Kbsgn;
   int ij, kl, ijkl;
@@ -184,9 +185,9 @@ void s1_block_vras(struct stringwr **alplist, struct stringwr **betlist,
                    int nlists, int nas, int nbs, int Ib_list, int Jb_list,
                    int Jb_list_nbs) {
   struct stringwr *Ib, *Kb;
-  unsigned int Ia_idx, Ib_idx, Kb_idx, Jb_idx;
-  unsigned int Ibcnt, Kbcnt, Kb_list, Ib_ex, Kb_ex;
-  unsigned int *Ibridx, *Kbridx;
+  size_t Ia_idx, Ib_idx, Kb_idx, Jb_idx;
+  size_t Ibcnt, Kbcnt, Kb_list, Ib_ex, Kb_ex;
+  size_t *Ibridx, *Kbridx;
   int *Ibij, *Kbij, *Iboij, *Kboij;
   signed char *Ibsgn, *Kbsgn;
   int ij, kl, ijkl, oij, okl;
