+++ psi4/src/psi4/detci/s2v.cc
@@ -3,23 +3,24 @@
  *
  * Psi4: an open-source quantum chemistry software package
  *
- * Copyright (c) 2007-2017 The Psi4 Developers.
+ * Copyright (c) 2007-2018 The Psi4 Developers.
  *
  * The copyrights for code used from other parties are included in
  * the corresponding files.
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
+ * This file is part of Psi4.
  *
- * This program is distributed in the hope that it will be useful,
+ * Psi4 is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * Psi4 is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with Psi4; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  *
  * @END LICENSE
@@ -70,9 +71,9 @@ void s2_block_vfci(struct stringwr **alplist, struct stringwr **betlist,
                    int nlists, int nas, int nbs, int Ia_list, int Ja_list,
                    int Ja_list_nas) {
   struct stringwr *Ia, *Ka;
-  unsigned int Ia_idx, Ib_idx, Ka_idx, Ja_idx;
-  unsigned int Iacnt, Kacnt, Ka_list, Ia_ex, Ka_ex;
-  unsigned int *Iaridx, *Karidx;
+  size_t Ia_idx, Ib_idx, Ka_idx, Ja_idx;
+  size_t Iacnt, Kacnt, Ka_list, Ia_ex, Ka_ex;
+  size_t *Iaridx, *Karidx;
   int *Iaij, *Kaij;
   signed char *Iasgn, *Kasgn;
   int ij, kl, ijkl;
@@ -172,9 +173,9 @@ void s2_block_vras(struct stringwr **alplist, struct stringwr **betlist,
                    int nlists, int nas, int nbs, int Ia_list, int Ja_list,
                    int Ja_list_nas) {
   struct stringwr *Ia, *Ka;
-  unsigned int Ia_idx, Ib_idx, Ka_idx, Ja_idx;
-  unsigned int Iacnt, Kacnt, Ka_list, Ia_ex, Ka_ex;
-  unsigned int *Iaridx, *Karidx;
+  size_t Ia_idx, Ib_idx, Ka_idx, Ja_idx;
+  size_t Iacnt, Kacnt, Ka_list, Ia_ex, Ka_ex;
+  size_t *Iaridx, *Karidx;
   int *Iaij, *Kaij, *Iaoij, *Kaoij;
   signed char *Iasgn, *Kasgn;
   int ij, kl, ijkl, oij, okl;
