+++ psi4/src/psi4/detci/s3v.cc
@@ -3,23 +3,24 @@
  *
  * Psi4: an open-source quantum chemistry software package
  *
- * Copyright (c) 2007-2017 The Psi4 Developers.
+ * Copyright (c) 2007-2018 The Psi4 Developers.
  *
  * The copyrights for code used from other parties are included in
  * the corresponding files.
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
+ * This file is part of Psi4.
  *
- * This program is distributed in the hope that it will be useful,
+ * Psi4 is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * Psi4 is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with Psi4; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  *
  * @END LICENSE
@@ -69,11 +70,11 @@ void s3_block_vdiag(struct stringwr *alplist, struct stringwr *betlist,
                     int Jb_sym, double **Cprime, double *F, double *V,
                     double *Sgn, int *L, int *R, int norbs, int *orbsym) {
   struct stringwr *Ia;
-  unsigned int Ia_ex;
+  size_t Ia_ex;
   int ij, i, j, t, kl, I, J, RJ;
   double tval, VS, *CprimeI0, *CI0;
   int jlen, Jacnt, *Iaij, Ia_idx;
-  unsigned int *Iaridx;
+  size_t *Iaridx;
   signed char *Iasgn;
   double *Tptr;
   int npthreads, rc, status;
@@ -149,11 +150,11 @@ void s3_block_v(struct stringwr *alplist, struct stringwr *betlist, double **C,
                 double **Cprime, double *F, double *V, double *Sgn, int *L,
                 int *R, int norbs, int *orbsym) {
   struct stringwr *Ia;
-  unsigned int Ia_ex;
+  size_t Ia_ex;
   int ij, i, j, kl, ijkl, I, J, RJ;
   double tval, VS, *CprimeI0, *CI0;
   int jlen, Ia_idx, Jacnt, *Iaij;
-  unsigned int *Iaridx;
+  size_t *Iaridx;
   signed char *Iasgn;
   double *Tptr;
 
@@ -223,7 +224,7 @@ int form_ilist(struct stringwr *alplist, int Ja_list, int nas, int kl, int *L,
   int inum = 0, Ia_idx, Ia_ex, Iacnt, ij;
   int *Iaij;
   struct stringwr *Ia;
-  unsigned int *Iaridx;
+  size_t *Iaridx;
   signed char *Iasgn;
 
   /* loop over Ia */
